feat(ME-101): Creating adminRoleOnCompanyCreated code
